Remove MapGuide/FDO submodules

+++ remove_submodules.sh
@@ -1,8 +1,25 @@
 #!/bin/sh
-git submodule deinit fdo
-git submodule deinit mapguide
-rm -rf fdo
-rm -rf mapguide
-rm -rf .git/modules/fdo
-rm -rf .git/modules/mapguide
-rm .gitmodules
+if [ -d fdo ];
+then
+    echo "De-init and remove FDO"
+    git submodule deinit -f fdo
+    rm -rf fdo
+fi
+if [ -d mapguide ];
+then
+    echo "De-init and remove MapGuide"
+    git submodule deinit -f mapguide
+    rm -rf mapguide
+fi
+if [ -d .git/modules/fdo ];
+then
+    rm -rf .git/modules/fdo
+fi
+if [ -d .git/modules/mapguide ];
+then
+    rm -rf .git/modules/mapguide
+fi
+if [ -f .gitmodules ];
+then
+    rm .gitmodules
+fi
